My otclient

+++ src/client/creature.h
@@ -57,7 +57,7 @@ public:
     void setOutfit(const Outfit& outfit);
     void setOutfitColor(const Color& color, int duration);
     void setLight(const Light& light) { m_light = light; }
-    void setSpeed(uint16 speed);
+    void setSpeed(double speed);
     void setBaseSpeed(double baseSpeed);
     void setSkull(uint8 skull);
     void setShield(uint8 shield);
@@ -84,7 +84,7 @@ public:
     Otc::Direction getDirection() { return m_direction; }
     Outfit getOutfit() { return m_outfit; }
     Light getLight() { return m_light; }
-    uint16 getSpeed() { return m_speed; }
+    double getSpeed() { return m_speed; }
     double getBaseSpeed() { return m_baseSpeed; }
     uint8 getSkull() { return m_skull; }
     uint8 getShield() { return m_shield; }
@@ -151,7 +151,7 @@ protected:
     Otc::Direction m_direction;
     Outfit m_outfit;
     Light m_light;
-    int m_speed;
+    double m_speed;
     double m_baseSpeed;
     uint8 m_skull;
     uint8 m_shield;
